% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd.R
\name{detect_binomial}
\alias{detect_binomial}
\alias{detect_logistic_regression}
\alias{fastcpd_binomial}
\alias{fastcpd.binomial}
\title{Find change points efficiently in logistic regression models}
\usage{
detect_binomial(data, ...)

detect_logistic_regression(data, ...)

fastcpd_binomial(data, ...)

fastcpd.binomial(data, ...)
}
\arguments{
\item{data}{A matrix or a data frame with the response variable as the first
column.}

\item{...}{Other arguments passed to \code{\link[=detect]{detect()}}, for example,
\code{segment_count}.}
}
\value{
A \link[=fastcpd-class]{fastcpd} object.
}
\description{
\code{\link[=detect_binomial]{detect_binomial()}} and \code{\link[=fastcpd.binomial]{fastcpd.binomial()}} are
wrapper functions of \code{\link[=detect]{detect()}} to find change points in
logistic regression models. The function is similar to \code{\link[=detect]{detect()}}
except that the data is by default a matrix or data frame with the response
variable as the first column and thus a formula is not required here.
}
\examples{
if (requireNamespace("mvtnorm", quietly = TRUE)) {
  set.seed(1)
  n <- 500
  p <- 4
  x <- mvtnorm::rmvnorm(n, rep(0, p), diag(p))
  theta <- rbind(rnorm(p, 0, 1), rnorm(p, 2, 1))
  y <- c(
    rbinom(300, 1, 1 / (1 + exp(-x[1:300, ] \%*\% theta[1, ]))),
    rbinom(200, 1, 1 / (1 + exp(-x[301:n, ] \%*\% theta[2, ])))
  )
  result <- suppressWarnings(detect_binomial(cbind(y, x)))
  summary(result)
  plot(result)
}
}
\seealso{
\code{\link[=detect]{detect()}}
}
