% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd.R
\name{detect_family}
\alias{detect_family}
\alias{fastcpd_family}
\alias{fastcpd.family}
\title{Family-specific change point detection}
\description{
Convenience functions built on \code{\link[=detect]{detect()}} for common models.
}
\seealso{
Basic statistics:
\code{\link[=detect_mean]{detect_mean()}}, \code{\link[=detect_variance]{detect_variance()}}, \code{\link[=detect_meanvariance]{detect_meanvariance()}},
\code{\link[=detect_exponential]{detect_exponential()}};

Regression:
\code{\link[=detect_lm]{detect_lm()}}, \code{\link[=detect_binomial]{detect_binomial()}}, \code{\link[=detect_poisson]{detect_poisson()}},
\code{\link[=detect_lasso]{detect_lasso()}}, \code{\link[=detect_quantile]{detect_quantile()}};

Time series:
\code{\link[=detect_ar]{detect_ar()}}, \code{\link[=detect_var]{detect_var()}}, \code{\link[=detect_arima]{detect_arima()}}, \code{\link[=detect_arma]{detect_arma()}},
\code{\link[=detect_garch]{detect_garch()}}, \code{\link[=detect_time_series]{detect_time_series()}};

Distribution-free:
\code{\link[=detect_kernel]{detect_kernel()}}, \code{\link[=detect_rank]{detect_rank()}}.
}
