% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd.R
\name{detect_garch}
\alias{detect_garch}
\alias{fastcpd_garch}
\alias{fastcpd.garch}
\title{Find change points efficiently in GARCH(\eqn{p}, \eqn{q}) models}
\usage{
detect_garch(data, order = c(0, 0), ...)

fastcpd_garch(data, order = c(0, 0), ...)

fastcpd.garch(data, order = c(0, 0), ...)
}
\arguments{
\item{data}{A numeric vector, a matrix, a data frame or a time series object.}

\item{order}{A positive integer vector of length two specifying the order of
the GARCH model.}

\item{...}{Other arguments passed to \code{\link[=detect]{detect()}}, for example,
\code{segment_count}.}
}
\value{
A \link[=fastcpd-class]{fastcpd} object.
}
\description{
\code{\link[=detect_garch]{detect_garch()}} and \code{\link[=fastcpd.garch]{fastcpd.garch()}} are
wrapper functions of \code{\link[=detect]{detect()}} to find change points in
GARCH(\eqn{p}, \eqn{q}) models. The function is similar to \code{\link[=detect]{detect()}}
except that the data is by default a one-column matrix or univariate vector
and thus a formula is not required here.
}
\examples{
\donttest{
set.seed(1)
n <- 401
sigma_2 <- rep(1, n + 1)
x <- rep(0, n + 1)
for (i in seq_len(200)) {
  sigma_2[i + 1] <- 20 + 0.8 * x[i]^2 + 0.1 * sigma_2[i]
  x[i + 1] <- rnorm(1, 0, sqrt(sigma_2[i + 1]))
}
for (i in 201:n) {
  sigma_2[i + 1] <- 1 + 0.1 * x[i]^2 + 0.5 * sigma_2[i]
  x[i + 1] <- rnorm(1, 0, sqrt(sigma_2[i + 1]))
}
result <- suppressWarnings(
  detect_garch(x[-1], c(1, 1), include.mean = FALSE)
)
summary(result)
plot(result)
}
\donttest{
set.seed(1)
n <- 120
sigma_2 <- rep(1, n + 1)
x <- rep(0, n + 1)
for (i in seq_len(60)) {
  sigma_2[i + 1] <- 10 + 0.5 * x[i]^2 + 0.3 * sigma_2[i]
  x[i + 1] <- rnorm(1, 0, sqrt(sigma_2[i + 1]))
}
for (i in 61:n) {
  sigma_2[i + 1] <- 0.2 + 0.05 * x[i]^2 + 0.1 * sigma_2[i]
  x[i + 1] <- rnorm(1, 0, sqrt(sigma_2[i + 1]))
}
result <- suppressWarnings(
  detect_garch(x[-1], c(1, 1), include.mean = FALSE)
)
summary(result)
plot(result)
}
\donttest{
set.seed(1)
n <- 150
sigma_2 <- rep(1, n + 1)
x <- rep(0, n + 1)
for (i in seq_len(75)) {
  sigma_2[i + 1] <- 10 + 0.5 * x[i]^2 + 0.3 * sigma_2[i]
  x[i + 1] <- rnorm(1, 0, sqrt(sigma_2[i + 1]))
}
for (i in 76:n) {
  sigma_2[i + 1] <- 0.2 + 0.05 * x[i]^2 + 0.1 * sigma_2[i]
  x[i + 1] <- rnorm(1, 0, sqrt(sigma_2[i + 1]))
}
result <- suppressWarnings(
  detect_garch(x[-1], c(1, 1), include.mean = FALSE)
)
summary(result)
plot(result)
}
}
\seealso{
\code{\link[=detect]{detect()}}
}
