% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd.R
\name{detect_mean}
\alias{detect_mean}
\alias{fastcpd_mean}
\alias{fastcpd.mean}
\title{Find change points efficiently in mean change models}
\usage{
detect_mean(data, ...)

fastcpd_mean(data, ...)

fastcpd.mean(data, ...)
}
\arguments{
\item{data}{A matrix, a data frame or a vector.}

\item{...}{Other arguments passed to \code{\link[=detect]{detect()}}, for example,
\code{segment_count}.}
}
\value{
A \link[=fastcpd-class]{fastcpd} object.
}
\description{
\code{\link[=detect_mean]{detect_mean()}} and \code{\link[=fastcpd.mean]{fastcpd.mean()}} are wrapper
functions of \code{\link[=detect]{detect()}} to find the mean change. The function is
similar to \code{\link[=detect]{detect()}} except that the data is by default a matrix or
data frame or a vector with each row / element as an observation and thus a
formula is not required here.
}
\examples{
set.seed(1)
data <- matrix(c(
  rnorm(300, mean = 0, sd = 10),
  rnorm(400, mean = 50, sd = 10),
  rnorm(300, mean = 2, sd = 10)
))
system.time(result <- detect_mean(data))
summary(result)
plot(result)
set.seed(1)
p <- 3
data <- rbind(
  matrix(rnorm(p * 3e+5, mean = 0, sd = 10), ncol = p),
  matrix(rnorm(p * 4e+5, mean = 50, sd = 10), ncol = p),
  matrix(rnorm(p * 3e+5, mean = 2, sd = 10), ncol = p)
)
system.time(result <- detect_mean(data, cp_only = TRUE))
summary(result)
}
\seealso{
\code{\link[=detect]{detect()}}
}
