% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd.R
\name{detect_variance}
\alias{detect_variance}
\alias{fastcpd_variance}
\alias{fastcpd.variance}
\title{Find change points efficiently in variance change models}
\usage{
detect_variance(data, ...)

fastcpd_variance(data, ...)

fastcpd.variance(data, ...)
}
\arguments{
\item{data}{A matrix, a data frame or a vector.}

\item{...}{Other arguments passed to \code{\link[=detect]{detect()}}, for example,
\code{segment_count}.}
}
\value{
A \link[=fastcpd-class]{fastcpd} object.
}
\description{
\code{\link[=detect_variance]{detect_variance()}} and \code{\link[=fastcpd.variance]{fastcpd.variance()}} are wrapper
functions of \code{\link[=detect]{detect()}} to find the variance change. The
function is similar to \code{\link[=detect]{detect()}} except that the data is by
default a matrix or data frame or a vector with each row / element as an
observation and thus a formula is not required here.
}
\examples{
set.seed(1)
data <- c(rnorm(300, 0, 1), rnorm(400, 0, 10), rnorm(300, 0, 1))
system.time(result <- detect_variance(data))
summary(result)
plot(result)
if (requireNamespace("mvtnorm", quietly = TRUE)) {
  set.seed(1)
  p <- 3
  data <- rbind(
    mvtnorm::rmvnorm(
      3e+5, rep(0, p), crossprod(matrix(runif(p^2) * 2 - 1, p))
    ),
    mvtnorm::rmvnorm(
      4e+5, rep(0, p), crossprod(matrix(runif(p^2) * 2 - 1, p))
    ),
    mvtnorm::rmvnorm(
      3e+5, rep(0, p), crossprod(matrix(runif(p^2) * 2 - 1, p))
    )
  )
  result_time <- system.time(
    result <- detect_variance(data, cp_only = TRUE)
  )
  print(result_time)
  summary(result)
}
}
\seealso{
\code{\link[=detect]{detect()}}
}
